requirements.txt
Click==7.0
Flask==1.0.2
flask-marshmallow==0.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
marshmallow-sqlalchemy==0.14.1
SQLAlchemy==1.2.12
Werkzeug==0.14.1
